requirements.txt
appdirs==1.4.4
asgiref==3.3.1
distlib==0.3.1
Django==3.1.7
django-crispy-forms==1.11.1
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
filelock==3.0.12
importlib-metadata==3.4.0
phonenumbers==8.12.18
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uWSGI==2.0.19.1
virtualenv==20.4.2
zipp==3.4.0
